Grove: merge Seeed updates
