--- PocketBeagle/.work-in-progress/Grove/Lesson-3.Switch-the-Music/README.md
@@ -1,30 +1,0 @@
-# Description:
-
-In this lesson, students will learn how to use the 2 buttons to select the next song or the last song. the buttons not only can select the next song or the last song by press instantly but also can play music by Press longly.
-
-The LCD will show the name of the song.
-
-## Hardware Requirement:
-
-- [Grove – Button x 2](http://wiki.seeedstudio.com/Grove-Button/)
-- [Grove – Speaker Plus](http://wiki.seeedstudio.com/Grove-Speaker/)
-- [Grove – 16x2 LCD](http://wiki.seeedstudio.com/Grove-16x2_LCD_Series/)
-
-## Hardware Connection
- 
-- Plug the Grove – Button into **A5** and **UART4** port
-- Plug the Grove – 16x2 LCD into **I2C1** port
-- Plug the Grove – Speaker Plus into **UART2** port
-- Power the Pocket Beagle via the **micro USB** port
-
-![](../img/project-3.jpg)
-
-
-## Software
-
-- Step 1. Enter Cloud9 by typing IP of Pocket Beagle
-- Step 2. Select PocketBeagle -> Grove -> Lesson-3.Switch-the-Music
-- Step 4. Run the Musical_Note.py by using Runner:Python.
-
-## success
-        Now please please slowly change the distance between your hand and the ultrasonic distance sensor, you can find the distance value in the LCD change and the music switched by the distance.
