--- PocketBeagle/.work-in-progress/Grove/Lesson-5.KeyBoard-Player/README.md
@@ -1,26 +1,0 @@
-# Description:
-
-In this lesson, students will learn how to use the capacitive touch sensor to play different musical note.
-
-## Hardware Requirement:
-
-- [Grove - 12 Key Capacitive I2C Touch Sensor V2](http://wiki.seeedstudio.com/Grove-12_Key_Capacitive_I2C_Touch_Sensor_V2-MPR121/)
-- [Grove – Speaker](http://wiki.seeedstudio.com/Grove-Speaker/)
-
-## Hardware Connection
-
-- Plug the Grove – Speaker into **UART2** port
-- Plug the Grove - 12 Key Capacitive I2C Touch Sensor V2 into **I2C2** port
-- Power the Pocket Beagle via the **micro USB** port
-
-![](../img/project-5.jpg)
-
-
-## Software
-
-- Step 1. Enter Cloud9 by typing IP of Pocket Beagle
-- Step 2. Select PocketBeagle -> Grove -> Lesson-5.KeyBoard-Player
-- Step 3. Run the KeyBoard_Player.py by using Runner:Python.
-
-## success
-        Try to touch the Capacitive key, image it as a keyboard, and play your music.
