--- PocketBeagle/.work-in-progress/Grove/Lesson-6.Start-the-Party/README.md
@@ -1,29 +1,0 @@
-# Description:
-
-In this lesson, students will learn how to use the capacitive touch sensor to play the song <Twinkle Star>. And the RGB LED will have different color based on different music note.
-
-### Hardware Requirement:
-
-- [Grove - 12 Key Capacitive I2C Touch Sensor V2](http://wiki.seeedstudio.com/Grove-12_Key_Capacitive_I2C_Touch_Sensor_V2-MPR121/)
-- [Grove - Speaker Plus](http://wiki.seeedstudio.com/Grove-Speaker/)
-- [Grove - Chainable RGB LED X 2](http://wiki.seeedstudio.com/Grove-Chainable_RGB_LED/)
-
-
-## Hardware Connection
-
-- Plug the Grove - Speaker Plus into **UART2** port
-- Plug the Grove - 12 Key Capacitive I2C Touch Sensor V2 into **I2C2** port
-- Plug the Grove - Chainable RGB LED X 2 into **A2** port
-- Power PocketBeagle via the **micro USB** port
-
-
-![](img/project-6.jpg)
-
-## Software
-
-- Step 1. Enter Cloud9 by typing IP of Pocket Beagle
-- Step 2. Select PocketBeagle -> Grove -> Lesson-6.Start-the-Party
-- Step 3. Run the Start_the_Party.py by using Runner:Python.
-
-## success
-        Just control the music as lesson 5, and you can see the color of RGB LED change.
