+++ PocketBeagle/.work-in-progress/Grove/tone_generator.py
@@ -4,8 +4,8 @@
 import numpy as np
 import pyaudio
 import wave
-import subprocess
-from tqdm import tqdm
+import os
+import time
 
 tone_freq_map={"do": 261.5, "re": 293.4,"me": 329.5,"fa": 349.1,"so": 391.7,"la": 440,"ti": 493.8,"do+":523}
 _SCALE_DEFS = [
@@ -20,7 +20,7 @@ _SCALE_DEFS = [
    ]
 channels = 1
 sampwidth = 2
-framerate = 44100
+framerate = 192000
 
 def Play_Music(file):
     # define stream chunk
@@ -51,14 +51,16 @@ def Play_Music(file):
     # close PyAudio
     p.terminate()
 
-def synthesizer(freq,duration = 0.5,amp=250,sampling_freq=framerate):
+def synthesizer(freq,duration = 10,amp=250,sampling_freq=framerate):
     samples = amp * (np.sin(2*np.pi*np.arange(sampling_freq*duration)*freq/sampling_freq))
     samples = samples.astype(np.float16)
     return samples
 
 def main():
-    # sudo alsactl restore 0 -f /opt/source/bb.org-overlays/extras/tlv320aic3104.state.txt
-    subprocess.call(['mkdir', '-p', '/tmp/scale'])
+    if not os.path.exists('/tmp/scale'):
+        os.popen('mkdir /tmp/scale')
+        while not os.path.exists('/tmp/scale'):
+            time.sleep(0.1)
     tone_freq = [ v for v in sorted(tone_freq_map.values())]
     for i in range(len(tone_freq)):
         f = wave.open( "/tmp/scale/%s"%_SCALE_DEFS[i],"wb")
@@ -67,7 +69,7 @@ def main():
         f.setframerate(framerate)
         f.writeframes(synthesizer(tone_freq[i]).tostring())
         f.close()
-        Play_Music("/tmp/scale/%s"%_SCALE_DEFS[i])
+        print("/tmp/scale/%s generated successfully"%_SCALE_DEFS[i])
 
 if __name__ == "__main__":
     main()
